+++ CHANGELOG.rst
@@ -4,6 +4,9 @@ CHANGELOG
 5.3.27 (unreleased)
 -------------------
 
+- Add `max_ops` property to `PatchField`, `BucketListField` and `BucketDictField`
+  [vangheem]
+
 - Add clear action to list, dict and annotation patch fields
   [vangheem]