+++ CHANGELOG.rst
@@ -4,6 +4,9 @@ CHANGELOG
 6.0.0a12 (unreleased)
 ---------------------
 
+- Import aiohttp only when recaptcha is configured
+  [masipcat]
+
 - Some asyncpg settings do not work with storages
   [vangheem]